+++ src/main.js
@@ -1,6 +1,21 @@
 import { createApp } from 'vue'
+import { createPinia } from 'pinia'
+import PrimeVue from 'primevue/config'
+import Aura from '@primevue/themes/aura'
+import router from '@/router/index.js'
 import App from './App.vue'
-import {router} from "@/router/index.js";
-import {createPinia} from "pinia";
 
-createApp(App).use(router).use(createPinia()).mount('#app')
+import 'primeicons/primeicons.css'
+import './assets/auth.css'
+
+const app = createApp(App)
+
+app.use(createPinia())
+app.use(router)
+app.use(PrimeVue, {
+    theme: {
+        preset: Aura
+    }
+})
+
+app.mount('#app')