+++ src/services/window.js
@@ -1,0 +1,3 @@
+export const GiveAlert = (msg) => {
+    alert(msg)
+}